+++ modules/tinymce/src/plugins/link/test/ts/browser/LinkDialogOverrideTest.ts
@@ -26,16 +26,16 @@ describe('browser.tinymce.plugins.link.LinkDialogOverrideTest', () => {
                   if (details.name === 'url' || details.name === 'link' || details.name === 'anchor') {
                     const data = api.getData();
                     if (data.url.value.length === 0) {
-                      api.disable('save');
+                      api.setEnabled('save', false);
                     } else {
-                      api.enable('save');
+                      api.setEnabled('save', true);
                     }
                   }
                 }
               });
               const api = originalWindowManager.open(newSpec);
               if (spec.initialData.url.value.length === 0) {
-                api.disable('save');
+                api.setEnabled('save', false);
               }
 
               return api;